+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 * Add read-only editor field support. Refs [UILD-630].
 * Enable repeatable subcomponents for all groups. Refs [UILD-632].
 * Add profile ID to generated record. Fixes [UILD-637].
+* Add ability to set profile as default. Refs [UILD-574].
 
 [UILD-552]:https://folio-org.atlassian.net/browse/UILD-552
 [UILD-544]:https://folio-org.atlassian.net/browse/UILD-544
@@ -71,6 +72,7 @@
 [UILD-630]:https://folio-org.atlassian.net/browse/UILD-630
 [UILD-632]:https://folio-org.atlassian.net/browse/UILD-632
 [UILD-637]:https://folio-org.atlassian.net/browse/UILD-637
+[UILD-574]:https://folio-org.atlassian.net/browse/UILD-574
 
 ## 1.0.5 (2025-04-30)
 * Fixed incorrect behavior when navigating between duplicated resources. Fixes [UILD-553].